changed lexer

+++ haflib/includes/garbage.h
@@ -6,7 +6,7 @@
 /*   By: conoel <conoel@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/07 17:15:57 by conoel            #+#    #+#             */
-/*   Updated: 2019/04/07 18:01:03 by conoel           ###   ########.fr       */
+/*   Updated: 2019/04/08 15:19:30 by conoel           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,6 +22,6 @@ typedef struct	s_malloc
 	void			*elem;
 }				t_malloc;
 
-void			*malloc_garbage(int size);
+void			*malloc_garbage(size_t size);
 
 #endif
